+++ lib/src/rust/frb_generated.io.dart
@@ -7,6 +7,7 @@ import 'api/ffi_api.dart';
 import 'dart:async';
 import 'dart:convert';
 import 'dart:ffi' as ffi;
+import 'error.dart';
 import 'frb_generated.dart';
 import 'markdown/ast.dart';
 import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated_io.dart';